refactoring and test fix
